Add --agent flag to Pi
Allows running TUI in the persona of a particular agent.

Fixes https://github.com/gordonbrander/busytown-pi/issues/26
